DEV-111: minor MR comment change
